C++ Project: Managing Variables and Constants
The purpose of this code is to accept variables and store the data. (For example, If a Professor needs to assign a grade to a student: "Alex = 92". After, the data will be recalled if the user asks for it by simply typing the Variable name and will also display a list with all the variables and the number of variables if the word "list" is typed.